requirements.txt
asgiref==3.7.2
Django==4.2.3
gunicorn==20.1.0
https://github.com/kkawabat/MagicEyeUtil/zipball/master
mistune==3.0.1
numpy==1.25.1
Pillow==10.0.0
psycopg2-binary==2.9.6
Pygments==2.15.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
